+++ src/vs/platform/update/electron-main/updateIpc.ts
