+++ src/vs/workbench/contrib/update/browser/media/markdown.css
